requirements.txt
accelerate==1.10.1
av>=12.0.0
bitsandbytes>=0.47.0
deepspeed==0.14.4
numpy==2.2.6
peft==0.17.1
pillow>=10.3.0
PyYAML==6.0.2
sentencepiece>=0.2.0
torch>=2.8.0
torchvision>=0.23.0
tqdm==4.67.1
transformers==4.45.2
wandb==0.22.0
